modules/legacy/MyAgents/projects/Agent-GCPtoolkit/pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "agent-gcptoolkit"
version = "0.1.1"
description = "GCP toolkit for agent operations"
requires-python = ">=3.10"
dependencies = [
    "google-cloud-secret-manager",
    "pyyaml",
]

[project.scripts]
gcptoolkit = "agent_gcptoolkit.cli.main:main"

[tool.setuptools]
packages.find = {where = ["."], exclude = ["tests", "tests.*", "build", "build.*"]}
